backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "regex>=2026.4.4",
    "sendgrid>=6.12.5",
    "uvicorn>=0.44.0",
]

[tool.uv]
package = true
